1. 매너점수

+++ backend/java-spring/src/main/java/com/example/gachi/model/Manner.java
@@ -24,8 +24,7 @@ public class Manner extends BaseEntity {
     @Enumerated(value = EnumType.STRING)
     private Review review;
 
-
     @Column(nullable = false)
-    private int score;
+    private long score;
 
 }
